refactor(main judge): move heat logic to hook
This file got very long and difficult to navigate. I have moved the
logic of creating the heat into a hook called "useHeatControls". It is
a complete copy-paste of the code, so no new functionality has been
added.
