+++ src/app/hooks/index.tsx
@@ -1,4 +1,5 @@
 import useFetchDataConvex from "./useFetchDataConvex";
+import useHeatControls from "./useHeatControls";
 import useYearSelector from "./useYearSelector";
 
-export { useFetchDataConvex, useYearSelector };
+export { useFetchDataConvex, useHeatControls, useYearSelector };
